requirements.txt
Flask==2.3.5
python-dotenv==1.0.1
authlib==1.2.1
supabase==1.0.2
requests==2.31.0
openai==1.8.4